requirements.txt
streamlit
pandas
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-api-python-client
openai
